Add deprecation notice to README.md
It has been a few years since the last update and jemalloc has had
several releases in that time. I've been maintaining my fork and
provided support on it for a couple years now and I think it's time to
deprecate this one.

This buildpack has had a great life (and should continue to work as is)
but I think it's worth it to steer folks toward a maintained fork. For
example, it was recommended for Ruby to skip the 4.x branch and to stick
to 3.6.0 (released in 2015), but our team has had really good luck with
5.1.0 and 5.2.0.

There was a [pull request](https://github.com/mojodna/heroku-buildpack-jemalloc/pull/18)
to merge the fork into the buildpack but given the risk it would expose
anyone using the current buildpack to I don't think it is a good path
forward. For more details see this comment:
https://github.com/mojodna/heroku-buildpack-jemalloc/pull/18#issuecomment-434357178